Fix portability issue related to size_t
Before, some literals would be casted to size_t using lu suffix,
assuming size_t is a long unsigned int, which isn't the case in all
platforms. The literal is now casted to size_t instead, ensuring
portability.

+++ src/StarSystem.cpp
@@ -84,7 +84,7 @@ StarSystem::StarSystem(std::string n) : name(n){
 						minDistance = dist;
 					}
 				}
-				distance = randomInt(60000000,std::min(200000000lu,minDistance/2));
+				distance = randomInt(60000000,std::min(size_t(200000000),minDistance/2));
 			}else{
 				distance = randomInt(60000000,200000000);
 			}
